Remove `include/grpcpp/impl/codegen/async_stream.h`
