+++ include/grpcpp/impl/codegen/async_generic_service.h
@@ -23,10 +23,10 @@
 
 #include <grpc/impl/codegen/port_platform.h>
 
-#include <grpcpp/impl/codegen/async_stream.h>
 #include <grpcpp/impl/codegen/byte_buffer.h>
 #include <grpcpp/impl/codegen/server_callback.h>
 #include <grpcpp/impl/codegen/server_callback_handlers.h>
+#include <grpcpp/support/async_stream.h>
 
 struct grpc_server;
