+++ test/cpp/codegen/compiler_test_mock_golden
@@ -5,8 +5,8 @@
 #include "src/proto/grpc/testing/compiler_test.pb.h"
 #include "src/proto/grpc/testing/compiler_test.grpc.pb.h"
 
-#include <grpcpp/impl/codegen/async_stream.h>
 #include <grpcpp/impl/codegen/sync_stream.h>
+#include <grpcpp/support/async_stream.h>
 #include <gmock/gmock.h>
 namespace grpc {
 namespace testing {
